celery_worker/requirements.txt
# worker
celery
flower
redis

# Tensorflow
tensorflow-cpu
pillow
opencv-python
matplotlib
